requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
groq==0.4.2
requests==2.31.0

